# atsamd5x: improve USBCDC