@sago35 2020-10-20
Since up to 128 bytes can be sent in a single pass, writing up to 128 bytes is very fast.
It is not very efficient when split into multiple Flush()'s, as there is a gap between Flush() and Flush().
However, it is much faster than the original implementation.

We can make it even faster by using the send completion interrupt.


* 10242 byte
  * old : 139.580 ms :  73377 B/s
  * new :  80.004 ms : 128018 B/s
* 128 byte
  * old : 1.711 ms :  74810 B/s
  * new : 0.154 ms : 831168 B/s
* 14 byte
  * old : 184.75  us :  75778 B/s
  * new :  17.41  us : 804135 B/s


testcode:

```
testdata := [10242]byte{}

machine.LED.High()
machine.UART0.Write(testdata[:])
machine.LED.Low()
```