@sago35 2020-10-20
You can increase the size of the RingBuffer by setting the following head and tail to volatile.Register16
If the size is increased, the amount of data that can be sent in one flush() will be larger and the transfer efficiency will improve.



```go
// src/machine/buffer.go
type RingBuffer struct {
	rxbuffer [bufferSize]volatile.Register8
	head     volatile.Register8
	tail     volatile.Register8
}
```

If you want to make it larger than 128, you must also change the following
Because the size of udd_ep_in_cache_buffer is 128.

```go
usbEndpointDescriptors[usb_CDC_ENDPOINT_IN].DeviceDescBank[1].ADDR.Set(uint32(uintptr(unsafe.Pointer(&udd_ep_in_cache_buffer[usb_CDC_ENDPOINT_IN]))))
```