@deadprogram 2020-10-20
@sago35 it appears you have implemented the interrupt solution:

```go
if i == usb_CDC_ENDPOINT_IN {
	UART0.waitTxc = false
}
```
However it appears that your proposed `udd_ep_in_cache_buffer` size increase is not in this PR.

Is that correct?