@deadprogram 2020-10-22
Also, it would seem that a similar change could be made to the SAMD21 support, is that true @sago35 ?