@sago35 2020-10-27
@deadprogram 

> However it appears that your proposed udd_ep_in_cache_buffer size increase is not in this PR.

I'll do a `git push` later.

> Also, it would seem that a similar change could be made to the SAMD21 support, is that true

Yes.
SAMD21 can be modified in the same way.

but,,,
There's a race condition that needs to be resolved for now.
I'd like to use `interrupt.Disable()`, but it's not currently available in the machine package.