@aykevl 2020-10-27
I'm not sure I fully understand this code, but I don't see anything weird with it.

Sidenote: this gives me even more reason that USB should be separated out in a new package. I think the code would be easier to follow if the code that sends a buffer over USB and the code that buffers for the next transmission is cleanly separated with an API.