@deadprogram 2020-10-28
@sago35 what is the specific race condition you are seeing?