@sago35 2020-11-02
> what is the specific race condition you are seeing?

Calling usbcdc.TxBuffer.Put(c) with multiple priorities will cause problems.

interrupt.Disable() can be built and used with interp-rewrite branch.
I believe #1430 will be merged with tinygo 0.17.0, so this PR will also be targeted at 0.17.0.

> Note that this should probably be merged after the next release, to give it some time to find possible bugs.
> https://github.com/tinygo-org/tinygo/pull/1430#issuecomment-718715299